requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
Deprecated==1.2.13
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==2.3.0
pymemcache==3.5.1
pyparsing==3.0.7
pytz==2022.1
redis==4.2.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
gunicorn==20.1.0
flower==1.0.0
watchdog==2.1.7
